move pages and their files to dedicated folders
