+++ src/pages/Game/css/game.css
