+++ src/pages/Game/scss/game.scss
